Major API change to version 0.9.0
Allows for custom channel implementations, to allow one-way channels, and one-to-many, or many-to-many type channels, or channels with custom filtering, for example.

Default ChannelFactory implementations can be found in the Channel interface.

Note that ChannelFilter (and related classes and fields) has been renamed to ChannelSelector to better reflect how the implementation actually works, as the old name has caused much confusion in the past.